+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+- `ManagerRegistryMock::getEntityManagerMock()` — change visibility from `private` to `protected` to allow override in subclasses
+- `ManagerRegistryMock::getClassMetadataMock()` — change visibility from `private` to `protected` to allow override in subclasses
+- `ManagerRegistryMock::getConnectionMock()` — change visibility from `private` to `protected` to allow override in subclasses
+
 ## [v3.1.1] - 2026-04-10
 
 ### Fixed